Modified the entire architecture of the program
Every protocol has a header which has all data that isn't lengthy like TCP options ARP addresses etc.
The Protocol classes like Ethernet class, is a class that handles every thing special about it like calculating checksum, set values, get values etc...
The protocol classes are a great thing for ease of creation and use, the program now tries to use no dynamic memory whatsoever.
There is a Mutable Packet and an IMMutable Packet. Both are from the Packet class which has the MTU of 1500 bytes, meaning it's not dynamic and wastes a bit of memory, but each packet has a reserved 1500 bytes.
The idea for immutable packets is that they will be on a special ring that is only to be read from, those are immutable packets that won't be modified in size but can be modified in their attributes.
All that is needed to do now is to ensure all protocols are working correctly and it is much easier to create packets with layers nicely.
An addition in this way is that now there are instead of multiple functions with a lot of copy pasting, now you have only 2 real functions that can be overriden (other than getSize()) : encodePre & encodePost, which you can add the checksum calculation and get other protocols that are near this one (useful for tcp getting IPv4 data)
It's a significant change in the right direction and hopefully now everything will look much better!